baremetal: launch the bootstrap VM with libvirt
Use libvirt to launch the bootstrap VM. Based on the bootstrap config
from libvirt platform driver.

The libvirt platform driver creates a base volume, and then creates
copy-on-write volumes for various VMs it creates. The baremetal
platform driver only creates a single libvirt VM, so it does not need
this additional layer.