baremetal: avoid separate base image for bootstrap
The libvirt platform driver creates a base volume, and then creates
copy-on-write volumes for various VMs it creates.

The baremetal platform driver only creates a single libvirt VM, so it
does not need this additional layer.

This fixes "destroy bootstrap" not cleaning up the base image. The
base image resource is owned by module.volume, but "destroy boostrap"
only destroys the resources in module.bootstrap. This makes sense for
the libvirt platform, but not for the baremetal platform.