bootstrap resolv.conf to use own coredns-mdns
In order for the clustering to work, the bootstrap node should be able
to resolve the SRV records for Etcd. In order to do that, the best is to
prepend localhost to /etc/resolv.conf so that as soon coredns-mdns is
ready, we will be able to use it to get the master entries.