baremetal: Dockerfile for installer image
This is just a variant of images/installer/Dockerfile.ci that builds
an image with kni-install at /usr/bin/openshift-install - this is so
'oc adm release extract --tools' knows how to extact it.

We also install libvirt-libs in the resulting image since the
openshift-install binary is dynamically linked when using libvirt
because it's built with CGO_ENABLED=1 and there is no static libvirt
library shipped in libvirt RPM builds.

Eventually, this dynamically linked binary will probably need to be
published as a separate image (or perhaps in installer-artifacts)
similarly to how the binary for mac is made available.