Work around extra host route.
This is the kni-installer version of the following commit:
https://github.com/openshift-metalkube/dev-scripts/commit/21cb51d7e2476824c8571fc73ebbdb003bd8df19

The dev-scripts commit fixed the issue for masters, but the issue can
still occur on the bootstrap VM, breaking the deployment process.

The issue is that there's a /32 route someimtes not getting deleted
when the VIP moves to another node.  This changes the keepalived
config such that the route created matches another route that is still
correct for the host.

Co-authored-by: Ben Nemec <bnemec@redhat.com>