Revert "Lower SRV record limit to 2"
This reverts commit b4351ad2c957c22343ad35c98244b7b01c94055b.

On bootstrap, we do not want to continue until we have all 3 nodes
available. Otherwise etcd will initially cluster with only 2 and
things get complicated. Since the DNS VIP moves off the bootstrap
once bootstrapping is complete, we should be okay to insist on 3
nodes here, but continue to allow 2 on the master coredns so that
the cluster remains operational if a node goes down.