Point Ignition at API VIP directly
OpenShift has moved to using api-int.domain as the DNS name for
internal API communication. This is problematic for baremetal because
we don't want to require users to configure an internal DNS name in
their external DNS, and we can't configure resolv.conf on the masters
to point at our coredns until after Ignition has run.

This change points Ignition directly at the API VIP specified by
the user in install-config. It only applies to baremetal - other
platforms will continue to use api-int.

Ref: https://bugzilla.redhat.com/show_bug.cgi?id=1685704