Lookup VIP from DNS if not provided during survey
We want to allow users to override the VIP that will be used for the
api-int DNS record, but to make the install as simple as possible we
should also be able to look it up from api.CLUSTER_DOMAIN if the
user is happy with that as the value.

Unfortunately we can't do the DNS lookup before asking the survey
question because it depends on the cluster domain config options,
and the cluster domain config option has a dependency on some
platform-specific configuration options so we can't get it before
asking platform-specific survey questions.

As an alternative, allow the user to leave the survey question
blank, which indicates that we should attempt to look it up from
DNS. If this lookup fails, we populate the config value with the
error message and add a validation for it. This will cause the install
to abort if the DNS lookup fails.