baremetal: Correct description of VRRP priority usage.
The baremetal VM actually has a higher VRRP priority than the control
plane nodes.  This ensures that the VIP does not move until the API is
fully functional on the control plane nodes.  The signal for this is
when the bootstrap VM is stopped.  That will leave only the lower
priority keepalived instances running that reside on the control plane
nodes.